+++ cdp-agentkit-core/cdp_agentkit_core/actions/__init__.py
@@ -8,7 +8,9 @@ from cdp_agentkit_core.actions.register_basename import RegisterBasenameAction
 from cdp_agentkit_core.actions.request_faucet_funds import RequestFaucetFundsAction
 from cdp_agentkit_core.actions.trade import TradeAction
 from cdp_agentkit_core.actions.transfer import TransferAction
+from cdp_agentkit_core.actions.wow.buy_token import WowBuyTokenAction
 from cdp_agentkit_core.actions.wow.create_token import WowCreateTokenAction
+from cdp_agentkit_core.actions.wow.sell_token import WowSellTokenAction
 
 
 # WARNING: All new CdpAction subclasses must be imported above, otherwise they will not be discovered
@@ -35,5 +37,7 @@ __all__ = [
     "TradeAction",
     "TransferAction",
     "WowCreateTokenAction",
+    "WowBuyTokenAction",
+    "WowSellTokenAction",
     "CDP_ACTIONS",
 ]